Move page components in separate folder
