+++ src/pages/AppBrowser/AppBrowser.css
