+++ src/pages/AppBrowser/AppBrowser.js
