+++ src/pages/AppBrowser/AppBrowser.vue
