+++ src/pages/Network.vue
