+++ src/pages/Setup/Setup.vue
