+++ src/pages/Transfer/Transfer.scss
